Create SCC before creating the deployment
With the existing sequence, the pod goes to crashloopback state since it can't bind the port. The reason is, that deployment is created prior to the SCC due to which, when the pod is created, we don't have the SCC created. If we just delete the pod, the next time pod comes up without any issues. If we create the SCC first, this works without any issues. Hence, moving the SCC template before creating the deployment object.